Use 32-bit launch ids in `PyLoadedExecutable`
`xla::ExecuteOptions::launch_id` is `int32_t`, so it makes sense to align `PyLoadedExecutable` with it. Since signed int overflow is UB in C++, the atomic that keeps track of the next launch id is defined as unsigned (so that overflow does not trigger UB), and then a cast is inserted on a newly minted launch id.

The current implementation may trigger UB because `options.launch_id = GetNextLaunchId()` implicitly casts `int64_t` to `int32_t`, causing signed int overflow.

PiperOrigin-RevId: 752916687

+++ jaxlib/py_executable.h
@@ -215,7 +215,7 @@ class PyLoadedExecutable {
   const ifrt::ExecuteOptions& options() const { return options_; }
 
   // Returns a unique launch ID to use for the next execution.
-  int64_t GetNextLaunchId();
+  int32_t GetNextLaunchId();
 
   const std::optional<std::string>& fingerprint() const { return fingerprint_; }
 
@@ -235,7 +235,7 @@ class PyLoadedExecutable {
   std::optional<std::string> fingerprint_;
 
   // Launch ID to use for the next execution.
-  std::atomic<int64_t> next_launch_id_;
+  std::atomic<uint32_t> next_launch_id_;
 
   // The options to pass to `executable_.Execute`.
   ifrt::ExecuteOptions options_;
